pyproject.toml
[tool.poetry]
name = "nonebot-plugin-naturel-gpt"
version = "2.2.0"
description = "一个基于NoneBot框架的Ai聊天插件，对接OpenAi文本生成接口"
authors = ["KroMiose <li_xiangff@163.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{ include = "nonebot_plugin_naturel_gpt" }]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc3"
openai = ">=0.27.0,<=0.28.0"
nonebot-adapter-onebot = "^2.2.1"
jieba = "^0.42.1"
numpy = "^1.21.3"
httpx = "^0.24.1"
aiosmtplib = "^2.0.1"
anyio = "^3.6.2"
aiohttp = "^3.8.4"
tiktoken = "0.4.0"

[tool.poetry.group.dev.dependencies]
nb-cli = "^1.2.8"
nonebot2 = {extras = ["fastapi"], version = "^2.1.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_naturel_gpt"]
plugin_dirs = []
builtin_plugins = []
